Fix for old behavior warning #29
This fixes below instances of the warning in system logs.

gnome-shell-ext: Some code called array.toString() on a Uint8Array instance. Previously this would have interpreted the bytes of the array as a string, but that is nonstandard. In the future this will return the bytes as comma-separated digits. For the time being, the old behavior has been preserved, but please fix your code anyway to explicitly call ByteArray.toString(array).
(Note that array.toString() may have been called implicitly.)
0 get_logs() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/lib.js":81]
1 buildPrefsWidget() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/prefs.js":364]
2 _selectExtension() ["resource:///org/gnome/shell/extensionPrefs/main.js":83]
3 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
4 _onCommandLine() ["resource:///org/gnome/shell/extensionPrefs/main.js":235]
5 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
6 wrapper() ["self-hosted":979]
7 main() ["resource:///org/gnome/shell/extensionPrefs/main.js":389]
8 <TOP LEVEL> ["<main>":1]

gnome-shell-ext: Some code called array.toString() on a Uint8Array instance. Previously this would have interpreted the bytes of the array as a string, but that is nonstandard. In the future this will return the bytes as comma-separated digits. For the time being, the old behavior has been preserved, but please fix your code anyway to explicitly call ByteArray.toString(array).
(Note that array.toString() may have been called implicitly.)
0 list_mouse_devices() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/lib.js":146]
1 search_touchpads() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/lib.js":177]
2 buildPrefsWidget() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/prefs.js":146]
3 _selectExtension() ["resource:///org/gnome/shell/extensionPrefs/main.js":83]
4 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
5 _onCommandLine() ["resource:///org/gnome/shell/extensionPrefs/main.js":235]
6 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
7 wrapper() ["self-hosted":979]
8 main() ["resource:///org/gnome/shell/extensionPrefs/main.js":389]
9 <TOP LEVEL> ["<main>":1]

gnome-shell-ext: Some code called array.toString() on a Uint8Array instance. Previously this would have interpreted the bytes of the array as a string, but that is nonstandard. In the future this will return the bytes as comma-separated digits. For the time being, the old behavior has been preserved, but please fix your code anyway to explicitly call ByteArray.toString(array).
(Note that array.toString() may have been called implicitly.)
0 _search_device() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":88]
1 _is_device() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":77]
2 _get_ids() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":51]
3 _init() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":40]
4 XInput() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":33]
5 <TOP LEVEL> ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/prefs.js":49]
6 _getExtensionPrefsModule() ["resource:///org/gnome/shell/extensionPrefs/main.js":67]
7 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
8 _selectExtension() ["resource:///org/gnome/shell/extensionPrefs/main.js":82]
9 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
10 _onCommandLine() ["resource:///org/gnome/shell/extensionPrefs/main.js":235]
11 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
12 wrapper() ["self-hosted":979]
13 main() ["resource:///org/gnome/shell/extensionPrefs/main.js":389]
14 <TOP LEVEL> ["<main>":1]

gnome-shell-ext: Some code called array.toString() on a Uint8Array instance. Previously this would have interpreted the bytes of the array as a string, but that is nonstandard. In the future this will return the bytes as comma-separated digits. For the time being, the old behavior has been preserved, but please fix your code anyway to explicitly call ByteArray.toString(array).
(Note that array.toString() may have been called implicitly.)
0 _get_all_ids() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":63]
1 _get_ids() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":49]
2 _init() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":40]
3 XInput() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":33]
4 <TOP LEVEL> ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/prefs.js":49]
5 _getExtensionPrefsModule() ["resource:///org/gnome/shell/extensionPrefs/main.js":67]
6 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
7 _selectExtension() ["resource:///org/gnome/shell/extensionPrefs/main.js":82]
8 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
9 _onCommandLine() ["resource:///org/gnome/shell/extensionPrefs/main.js":235]
10 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
11 wrapper() ["self-hosted":979]
12 main() ["resource:///org/gnome/shell/extensionPrefs/main.js":389]
13 <TOP LEVEL> ["<main>":1]

gnome-shell: Some code called array.toString() on a Uint8Array instance. Previously this would have interpreted the bytes of the array as a string, but that is nonstandard. In the future this will return the bytes as comma-separated digits. For the time being, the old behavior has been preserved, but please fix your code anyway to explicitly call ByteArray.toString(array).
(Note that array.toString() may have been called implicitly.)
0 _is_device_enabled() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":150]
1 _all_devices_enabled() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":169]
2 _switch_all_devices() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/xinput.js":143]
3 _touchpad_changed() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/extension.js":668]
4 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
5 wrapper() ["self-hosted":979]
6 _switch_touchpad() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/extension.js":648]
7 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
8 _onChangeSwitchMethod() ["/home/ashesh/.local/share/gnome-shell/extensions/touchpad-indicator@orangeshirt/extension.js":469]
9 wrapper() ["resource:///org/gnome/gjs/modules/_legacy.js":82]
10 wrapper() ["self-hosted":979]